requirements.txt
# requirements.txt

fastapi
uvicorn
pydantic
pandas
scikit-learn
joblib
minio
kafka-python==2.0.2  # Pastikan versi ini tetap
streamlit
plotly  # <- TAMBAHKAN INI